init resource 구현부 분리 && globla uuid_str 선언
