+++ keti-redfish/include/resource.hpp
@@ -14,6 +14,7 @@
  * @brief Open data protocol information
  */
 #define ODATA_TYPE_VERSION REDFISH_VERSION "_0_0"
+
 // Open data protocol path
 #define ODATA_SERVICE_ROOT_ID "/redfish/" REDFISH_VERSION
 #define ODATA_SYSTEM_ID ODATA_SERVICE_ROOT_ID "/Systems"
@@ -23,7 +24,6 @@
 #define ODATA_ETHERNET_INTERFACE_ID ODATA_MANAGER_ID "/EthernetInterfaces" 
 #define ODATA_TASK_SERVICE_ID ODATA_SERVICE_ROOT_ID "/TaskService"
 #define ODATA_TASK_ID ODATA_TASK_SERVICE_ID "/Tasks" 
-
 #define ODATA_SESSION_SERVICE_ID ODATA_SERVICE_ROOT_ID "/SessionService"
 #define ODATA_SESSION_ID ODATA_SESSION_SERVICE_ID "/Sessions"
 #define ODATA_ACCOUNT_SERVICE_ID ODATA_SERVICE_ROOT_ID "/AccountService"
@@ -63,10 +63,6 @@
 #define ODATA_MANAGER_COLLECTION_TYPE "#ManagerCollection.ManagerCollection"
 #define ODATA_MANAGER_TYPE "#Manager." ODATA_TYPE_VERSION ".Manager"
 #define ODATA_NETWORK_PROTOCOL_TYPE "#NetworkProtocol." ODATA_TYPE_VERSION ".NetworkProtocol" 
-
-#define ODATA_ACTIONS_TYPE "#Actions." ODATA_TYPE_VERSION ".Actions"
-#define ODATA_ACTIONS_COLLECTION_TYPE "#ActionsCollection.ActionsCollection"
-
 #define ODATA_ETHERNET_INTERFACE_COLLECTION_TYPE "#EthernetInterfaceCollection.EthernetInterfaceCollection"
 #define ODATA_ETHERNET_INTERFACE_TYPE "#EthernetInterface." ODATA_TYPE_VERSION ".EthernetInterface"
 #define ODATA_LOG_SERVICE_COLLECTION_TYPE "#LogServiceCollection.LogServiceCollection"
@@ -75,9 +71,7 @@
 #define ODATA_LOG_ENRTY_TYPE "#LogEntry." ODATA_TYPE_VERSION ".LogEntry"
 #define ODATA_TASK_SERVICE_TYPE "#TaskService." ODATA_TYPE_VERSION ".TaskService"
 #define ODATA_TASK_COLLECTION_TYPE "#TaskCollection.TaskCollection"
-
 #define ODATA_TASK_TYPE "#Task." ODATA_TYPE_VERSION ".Task" 
-
 #define ODATA_SESSION_SERVICE_TYPE "#SessionService." ODATA_TYPE_VERSION ".SessionService"
 #define ODATA_SESSION_COLLECTION_TYPE "#SessionCollection.SessionCollection"
 #define ODATA_SESSION_TYPE "#Session." ODATA_TYPE_VERSION ".Session"
@@ -626,6 +620,8 @@ public:
  */
 extern unordered_map<string, Resource *> g_record;
 extern map<string, string> module_id_table;
+extern string uuid_str;
+
 // extern unordered_map<string, unordered_map<string, Task *> > task_map;
 
 /**
@@ -822,7 +818,6 @@ public:
     bool load_json(json::value &j);
 };
 
-
 /**
  * @brief Redfish resource of Log Service
  * @authors 강
@@ -912,8 +907,8 @@ class LogService : public Resource
 };
 
 /**
- * @brief Redfish resource of Event Service
- * @authors 강
+ * @brief Redfish resource of Event
+ * @authors 김
  */
 class Event
 {
@@ -1014,7 +1009,6 @@ class EventService : public Resource
  * @brief Redfish resource of Update Service
  * @authors 강
  */
-
 class SoftwareInventory : public Resource
 {
     public:
@@ -1096,11 +1090,10 @@ class UpdateService : public Resource
     bool SimpleUpdate(json::value body);
 };
 
-// /**
-//  * @brief Redfish resource of Certificate
-//  * @authors 김
-//  */
-
+/**
+ * @brief Redfish resource of Certificate
+ * @authors 김
+ */
 class Certificate : public Resource
 {
     public:
@@ -1120,26 +1113,6 @@ class Certificate : public Resource
 
     Certificate(const string _odata_id) : Resource(CERTIFICATE_TYPE, _odata_id, ODATA_CERTIFICATE_TYPE)
     {
-        this->id = "";
-        this->certificateString = "";
-        this->certificateType = "";
-
-        this->issuer.city = "";
-        this->issuer.commonName = "";
-        this->issuer.country = "";
-        this->issuer.email = "";
-        this->issuer.organization = "";
-        this->issuer.organizationUnit = "";
-        this->issuer.state = "";
-
-        this->subject.city = "";
-        this->subject.commonName = "";
-        this->subject.country = "";
-        this->subject.email = "";
-        this->subject.organization = "";
-        this->subject.organizationUnit = "";
-        this->subject.state = "";
-
         Actions rekey;
         rekey.type = RE_KEY;
         rekey.name = "#Certificate.Rekey";
@@ -1277,7 +1250,6 @@ class CertificateService : public Resource
  * @authors 강
  * @details 
  */
-
 class NetworkProtocol : public Resource
 {
     public:
@@ -1285,6 +1257,7 @@ class NetworkProtocol : public Resource
     string hostname;
     string description;
     string fqdn;
+    string name;
     bool snmp_enabled;
     bool ipmi_enabled;
     bool ntp_enabled;
@@ -1334,6 +1307,7 @@ class EthernetInterfaces : public Resource
 {
 public:
     string id;
+    string name;
     string description;
     string link_status;
     string permanent_mac_address;
@@ -1376,6 +1350,7 @@ class Manager : public Resource
 {
 public:
     string id;
+    string name;
     string manager_type;
     string description;
     string uuid;
@@ -1429,7 +1404,7 @@ public:
         g_record.erase(this->odata.id);
     };
 
-    json::value get_json(void); // 여기 가서 수정해줘야함
+    json::value get_json(void); 
     bool load_json(json::value &j);
     bool Reset(json::value body);
 };
@@ -1579,8 +1554,6 @@ public:
     json::value get_json(void);
     bool load_json(json::value &j);
     pplx::task<void> start(void);
-
-private:
 };
 
 /**
@@ -1606,23 +1579,6 @@ public:
     // Class constructor, destructor oveloading
     Temperature(const string _odata_id) : Resource(TEMPERATURE_TYPE, _odata_id, ODATA_THERMAL_TYPE)
     {
-        // this->member_id = "";
-        // this->status.state = STATUS_STATE_ENABLED;
-        // this->status.health = STATUS_HEALTH_OK;
-
-        // this->sensor_num = 0;
-        // this->reading_celsius = 0;
-        // this->upper_threshold_non_critical = 0;
-        // this->upper_threshold_critical = 0;
-        // this->upper_threshold_fatal = 0;
-        // this->lower_threshold_non_critical = 0;
-        // this->lower_threshold_critical = 0;
-        // this->lower_threshold_fatal = 0;
-        // this->min_reading_range_temp = 0;
-        // this->max_reading_range_temp = 0;
-        // this->physical_context = "CPU";
-        // this->thread = false;
-
         g_record[_odata_id] = this;
     }
     Temperature(const string _odata_id, const string _member_id) : Temperature(_odata_id)
@@ -1646,7 +1602,6 @@ private:
 /**
  * @brief Redfish resource of temperature
  */
-
 class Sensor : public Resource
 {
     public:
@@ -1668,28 +1623,6 @@ class Sensor : public Resource
 
     Sensor(const string _odata_id) : Resource(SENSOR_TYPE, _odata_id, ODATA_SENSOR_TYPE)
     {
-        // this->id = "";
-        // this->reading_type = "Such as Temperature";
-        // this->reading_time = "Reading Time";
-        // this->reading = 30.6;
-
-        // this->reading_units = "C";
-        // this->reading_range_min = 0;
-        // this->reading_range_max = 70;
-        // this->accuracy = 0.25;
-        // this->precision = 1;
-        // this->sensing_interval = "PT3S ???";
-        // this->physical_context = "Chassis";
-        // this->thresh.upper_caution.activation = "Increasing";
-        // this->thresh.upper_caution.reading = 35;
-        // this->thresh.upper_critical.activation = "Increasing";
-        // this->thresh.upper_critical.reading = 40;
-        // this->thresh.lower_caution.activation = "Increasing";
-        // this->thresh.lower_caution.reading = 10;
-
-        // this->status.state = STATUS_STATE_ENABLED;
-        // this->status.health = STATUS_HEALTH_OK;
-
         g_record[_odata_id] = this;
     }
     Sensor(const string _odata_id, const string _sensor_id) : Sensor(_odata_id)
@@ -1705,14 +1638,12 @@ class Sensor : public Resource
     bool load_json(json::value &j);
 };
 
-
 class Fan : public Resource
 {
 public:
     string member_id;
     Status status;
     int sensor_num;
-    
     int reading;
     string reading_units;
     
@@ -1729,25 +1660,6 @@ public:
     // Class constructor, destructor oveloading
     Fan(const string _odata_id) : Resource(FAN_TYPE, _odata_id)
     {
-        // this->member_id = "";
-        // this->status.state = STATUS_STATE_ENABLED;
-        // this->status.health = STATUS_HEALTH_OK;
-
-        // this->sensor_num = 0;
-        // this->reading = 0;
-        // this->reading_units = "RPM";
-
-        // this->lower_threshold_fatal = 0;
-        // this->lower_threshold_critical = 0;
-        // this->lower_threshold_non_critical = 0;
-        // this->upper_threshold_fatal = 0;
-        // this->upper_threshold_critical = 0;
-        // this->upper_threshold_non_critical = 0;
-
-        // this->min_reading_range = 0;
-        // this->max_reading_range = 0;
-        // this->physical_context = "CPU";
-
         g_record[_odata_id] = this;
     }
     Fan(const string _odata_id, const string _member_id) : Fan(_odata_id)
@@ -1781,14 +1693,6 @@ public:
         this->temperatures = nullptr;
         this->fans = nullptr;
 
-        // // Temperatures configuration
-        // this->temperatures = new List(this->odata.id + "/Temperatures", TEMPERATURE_TYPE);
-        // this->temperatures->name = "Chassis Temperatures";
-
-        // // Fans configuration
-        // this->fans = new List(this->odata.id + "/Fans", FAN_TYPE);
-        // this->fans->name = "Chassis Fans";
-
         g_record[_odata_id] = this;
     };
     ~Thermal()
@@ -1804,7 +1708,6 @@ public:
  * @brief Power 관련 resource
  * @authors 강
  */
-
 class Voltage : public Resource
 {
     public:
@@ -1825,22 +1728,6 @@ class Voltage : public Resource
 
     Voltage(const string _odata_id) : Resource(VOLTAGE_TYPE, _odata_id, ODATA_POWER_TYPE)
     {
-        // this->member_id = "";
-        // this->status.state = STATUS_STATE_ENABLED;
-        // this->status.health = STATUS_HEALTH_OK;
-
-        // this->sensor_num = 11;
-        // this->reading_volts = 12;
-        // this->upper_threshold_non_critical = 12.5;
-        // this->upper_threshold_critical = 13;
-        // this->upper_threshold_fatal = 15;
-        // this->lower_threshold_non_critical = 11.5;
-        // this->lower_threshold_critical = 11;
-        // this->lower_threshold_fatal = 10;
-        // this->min_reading_range = 0;
-        // this->max_reading_range = 20;
-        // this->physical_context = "VoltageRegulator";
-
         g_record[_odata_id] = this;
 
     }
@@ -1880,38 +1767,7 @@ class PowerSupply : public Resource
 
     PowerSupply(const string _odata_id) : Resource(POWER_SUPPLY_TYPE, _odata_id, ODATA_POWER_TYPE)
     {
-        // this->member_id = "";
-        // this->status.state = STATUS_STATE_ENABLED;
-        // this->status.health = STATUS_HEALTH_OK;
-        
-        this->power_supply_type = "AC";
-        this->line_input_voltage_type = "ACWideRange";
-        this->line_input_voltage = 120;
-        this->power_capacity_watts = 800;
-        this->last_power_output_watts = 325;
-
-        this->model = "";
-        this->manufacturer = "";
-        this->firmware_version = "";
-        this->serial_number = "";
-        this->part_number = "";
-        this->spare_part_number = "";
-
-        // InputRange input;
-        // input.input_type = "AC";
-        // input.minimum_voltage = 100;
-        // input.maximum_voltage = 120;
-        // input.output_wattage = 800;
-        // this->input_ranges.push_back(input);
-
-        // input.input_type = "AC";
-        // input.minimum_voltage = 200;
-        // input.maximum_voltage = 240;
-        // input.output_wattage = 1300;
-        // this->input_ranges.push_back(input);
-
         g_record[_odata_id] = this;
-
     }
     PowerSupply(const string _odata_id, const string _member_id) : PowerSupply(_odata_id)
     {
@@ -1943,24 +1799,6 @@ class PowerControl : public Resource
 
     PowerControl(const string _odata_id) : Resource(POWER_CONTROL_TYPE, _odata_id, ODATA_POWER_TYPE)
     {
-        // this->member_id = "";
-        // this->status.state = STATUS_STATE_ENABLED;
-        // this->status.health = STATUS_HEALTH_OK;
-        // this->power_consumed_watts = 0;
-        // this->power_requested_watts = 0;
-        // this->power_available_watts = 0;
-        // this->power_capacity_watts = 0;
-        // this->power_allocated_watts = 0;
-
-        // this->power_limit.limit_in_watts = 0;
-        // this->power_limit.limit_exception = "LogEventOnly";
-        // this->power_limit.correction_in_ms = 0;
-
-        // this->power_metrics.interval_in_min = 0;
-        // this->power_metrics.min_consumed_watts = 0;
-        // this->power_metrics.max_consumed_watts = 0;
-        // this->power_metrics.avg_consumed_watts = 0;
-
         g_record[_odata_id] = this;
     }
     PowerControl(const string _odata_id, const string _member_id) : PowerControl(_odata_id)
@@ -1992,19 +1830,6 @@ class Power : public Resource
         this->voltages = nullptr;
         this->power_supplies = nullptr;
 
-        // this->power_control = new List(this->odata.id + "/PowerControl", POWER_CONTROL_TYPE);
-        // this->power_control->name = "PowerControl";
-
-        // this->voltages = new List(this->odata.id + "/Voltages", VOLTAGE_TYPE);
-        // this->voltages->name = "Voltages";
-
-        // this->power_supplies = new List(this->odata.id + "/PowerSupplies", POWER_SUPPLY_TYPE);
-        // this->power_supplies->name = "PowerSupplies";
-
-        this->power_control = nullptr;
-        this->voltages = nullptr;
-        this->power_supplies = nullptr;
-
         g_record[_odata_id] = this;
     };
     ~Power()
@@ -2017,13 +1842,11 @@ class Power : public Resource
 
 };
 
-
 /**
  * @brief 시스템 리소스에 필요한 storage, bios, simplestorage, processors..  processorsummary,networkinterface는뺌
  * @authors 강
  * 
  */
-
 class Bios : public Resource
 {
     public:
@@ -2110,7 +1933,6 @@ class SimpleStorage : public Resource
 
     json::value get_json(void);
     bool load_json(json::value &j);
-
 };
 
 class StorageControllers : public Resource
@@ -2360,7 +2182,6 @@ class Processors : public Resource
 
     json::value get_json(void);
     bool load_json(json::value &j);
-
 };
 
 class Memory : public Resource
@@ -2395,14 +2216,11 @@ class Memory : public Resource
 
     json::value get_json(void);
     bool load_json(json::value &j);
-
 };
 
 /**
  * @brief Redfish resource of Systems
- * 
  */
-
 class Systems : public Resource
 {
     public :
@@ -2421,10 +2239,7 @@ class Systems : public Resource
     string power_state;
     uint8_t indicator_led;
     string bios_version;
-
-    Status status;
     string uuid;
-    Boot boot;
     // Ipmifru *fru_this;
     // Location location;
     // Thermal *thermal;
@@ -2432,6 +2247,8 @@ class Systems : public Resource
     // ProcessorSummary *ps; // 구조체로 바꿔야할듯 현재리소슨데
     // MemorySummary ms;
     Bios *bios; // resource Bios
+    Status status;
+    Boot boot;
     
     // Collection *network; // resource NetworkInterfaces // 일단 없음
     Collection *storage; // resource Storages
@@ -2648,7 +2465,7 @@ public:
         this->id = "RootService";
         this->name = "Root Service";
         this->redfish_version = "1.0.0";
-        this->uuid = "";
+        this->uuid = uuid_str;
 
         // CMM ID와 주소 등록
         if(module_id_table.find(CMM_ID) == module_id_table.end()) // 없으면 등록
@@ -2665,7 +2482,7 @@ public:
         
         // Collection Generate in ServiceRoot
         if (!record_is_exist(ODATA_SYSTEM_ID)){
-            log(info) << "System init";
+            log(info) << "[...]System init";
             system_collection = new Collection(ODATA_SYSTEM_ID, ODATA_SYSTEM_COLLECTION_TYPE);
             system_collection->name = "Computer System Collection";
 
@@ -2673,7 +2490,7 @@ public:
         }
 
         if (!record_is_exist(ODATA_CHASSIS_ID)){
-            log(info) << "Chassis init";
+            log(info) << "[...]Chassis init";
             chassis_collection = new Collection(ODATA_CHASSIS_ID, ODATA_CHASSIS_COLLECTION_TYPE);
             chassis_collection->name = "Chassis Collection";
         
@@ -2681,7 +2498,7 @@ public:
         }
 
         if (!record_is_exist(ODATA_MANAGER_ID)){
-            log(info) << "Manager init";
+            log(info) << "[...]Manager init";
             manager_collection = new Collection(ODATA_MANAGER_ID, ODATA_MANAGER_COLLECTION_TYPE);
             manager_collection->name = "Manager Collection";
 
@@ -2690,14 +2507,14 @@ public:
         
         // UpdateService configuration
         if (!record_is_exist(ODATA_UPDATE_SERVICE_ID)){
-            log(info) << "Update Service init";
+            log(info) << "[...]Update Service init";
             update_service = new UpdateService(ODATA_UPDATE_SERVICE_ID);
             
             init_update_service(update_service);
         }
         // TaskService configuration
         if (!record_is_exist(ODATA_TASK_SERVICE_ID)){
-            log(info) << "Task Service init";
+            log(info) << "[...]Task Service init";
             task_service = new TaskService();
         
             init_task_service(task_service);
@@ -2705,27 +2522,27 @@ public:
 
         // EventService configuration
         if (!record_is_exist(ODATA_EVENT_SERVICE_ID)){
-            log(info) << "Event Service init";
+            log(info) << "[...]Event Service init";
             event_service = new EventService();
 
             init_event_service(event_service);
         }
         if (!record_is_exist(ODATA_CERTIFICATE_SERVICE_ID)){
-            log(info) << "Certificate Service init";
+            log(info) << "[...]Certificate Service init";
             certificate_service = new CertificateService();
             certificate_service->certificate_location = new CertificateLocation();
         }
         
         // AccountService configuration
         if (!record_is_exist(ODATA_ACCOUNT_SERVICE_ID)){
-            log(info) << "Account Service init";
+            log(info) << "[...]Account Service init";
             account_service = new AccountService();
 
             init_account_service(account_service);
         }
         // SessionService configuration
         if (!record_is_exist(ODATA_SESSION_SERVICE_ID)){
-            log(info) << "Session Service init";
+            log(info) << "[...]Session Service init";
             session_service = new SessionService();
 
             init_session_service(session_service);
